# oapi/api.yml
openapi: 3.1.0
info:
  title: todo-app API
  description: |-
    ## 概要
    - todo-appのAPI仕様書
servers:
  - url: http://localhost:8083
    description: ローカル開発環境
...
paths:
...
  /categories:
    get:
      summary: 特定のボード内のカテゴリ一覧を取得するエンドポイント
      description: |-
        ## 概要
        - 特定のボード内のカテゴリ一覧を取得する
      tags:
        - categories
      operationId: getCategories
      security: []
      parameters:
        - name: board_id
          in: query
          description: カテゴリ一覧を取得するボードのIDをクエリパラメータで指定
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "200":
          description: 特定のボード内のカテゴリを全て返す
          content:
            application/json:
              schema:
                $ref: "./responses/categories.yml"
        default:
          $ref: "./errors/unexpected.yml"
    post:
      summary: カテゴリを新規作成するエンドポイント
      description: |-
        ## 概要
        - カテゴリのタイトルを受け取り、新規作成したカテゴリ情報を返す
      tags:
        - categories
      operationId: createCategories
      security: []
      requestBody:
        description: 新規作成するカテゴリ
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                board_id:
                  type: integer
                  description: 新規作成するカテゴリのボードID
                  example: 1
                category_title:
                  type: string
                  description: 新規作成するカテゴリのタイトル
                  example: TODO
              required:
                - board_id
                - category_title
      responses:
        "201":
          description: Successfully create a new category
          content:
            application/json:
              schema:
                $ref: "./entity/category.yml"
        default:
          $ref: "./errors/unexpected.yml"
  /categories/{category_id}:
    post:
      summary: カテゴリタイトルを更新するエンドポイント
      description: |-
        ## 概要
        - カテゴリのタイトルを受け取り、更新したカテゴリ情報を返す
      tags:
        - categories
      operationId: updateCategory
      security: []
      parameters:
        - name: category_id
          in: path
          description: 更新するカテゴリのID
          required: true
          schema:
            type: integer
            example: 1
      requestBody:
        description: 更新するカテゴリ
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                category_title:
                  type: string
                  description: 更新するカテゴリのタイトル
                  example: TODO
              required:
                - category_title
      responses:
        "200":
          description: Successfully update a category
          content:
            application/json:
              schema:
                $ref: "./entity/category.yml"
        default:
          $ref: "./errors/unexpected.yml"
    delete:
      summary: カテゴリを削除するエンドポイント
      description: |-
        ## 概要
        - カテゴリIDを受け取り、カテゴリを削除する
      tags:
        - categories
      operationId: deleteCategory
      security: []
      parameters:
        - name: category_id
          in: path
          description: 削除するカテゴリのID
          required: true
          schema:
            type: integer
            example: 1
      responses:
        "204":
          description: Successfully delete a category
        default:
          $ref: "./errors/unexpected.yml"
...